refactor(layout): merge AccessoriesShell + CompanyShell into SidebarShell
Both shells rendered the same lt-wrap > Breadcrumbs > grid layout with
side nav + main panel, differing only in the CSS namespace ("acc" vs
"co") and which content/side-nav was wired in. Replaces both with a
single generic SidebarShell that takes the side nav as a slot prop and
the wrapper/main class names as strings.

Callers (accessories + company pages) now own the content lookup and
side-nav rendering inline, plus their domain CSS import.

Net: -44 LOC.
